+++ tests/routes/ignore.js
@@ -1,0 +1,11 @@
+export default {
+  ignoredGet: {
+    method: 'GET',
+    path: '/ignored',
+    config: {
+      async handler(req, res) {
+        res.send('You should not be getting this message');
+      },
+    },
+  },
+};